--- ToDo-Garden/ToDo-Garden/ManageGroupScene/Sources/ManageGroupScene/DeleteGroupWorker.swift
@@ -1,18 +1,0 @@
-//
-//  DeleteGroupWorker.swift
-//
-//
-//  Created by SONG on 7/2/24.
-//
-
-import Foundation
-
-import ManageGroupSceneAPI
-import ManageGroupSceneEntity
-
-public struct DeleteGroupWorker: DeleteGroupWorkable {
-  public init() {}
-  public func deleteGroup(request: ManageGroup.DeleteGroup.Request) {
- 
-  }
-}
