--- ToDo-Garden/ToDo-Garden/ManageGroupScene/Sources/ManageGroupScene/FetchGroupListWorker.swift
@@ -1,18 +1,0 @@
-//
-//  ManageGroupWorker.swift
-//  
-//
-//  Created by SONG on 6/26/24.
-//  Copyright (c) 2024 ToDoGarden. All rights reserved.
-
-import Foundation
-
-import ManageGroupSceneAPI
-import ManageGroupSceneEntity
-
-public struct FetchGroupListWorker: FetchGroupListWorkable {
-  public init() {}
-  public func fetchGroupList(request: ManageGroup.FetchGroupList.Request) {
-    
-  }
-}
