--- ToDo-Garden/ToDo-Garden/ManageGroupScene/Sources/ManageGroupScene/ReorderGroupWorker.swift
@@ -1,18 +1,0 @@
-//
-//  ReorderGroupWorker.swift
-//
-//
-//  Created by SONG on 7/2/24.
-//
-
-import Foundation
-
-import ManageGroupSceneAPI
-import ManageGroupSceneEntity
-
-public struct ReorderGroupWorker: ReorderGroupWorkable {
-  public init() {}
-  public func reorderGroup(request: ManageGroup.ReorderGroup.Request) {
-    
-  }
-}
