--- ToDo-Garden/ToDo-Garden/ManageGroupScene/Sources/ManageGroupSceneAPI/DeleteGroupWorkable.swift
@@ -1,14 +1,0 @@
-//
-//  DeleteGroupWorkable.swift
-//
-//
-//  Created by SONG on 7/1/24.
-//
-
-import Foundation
-
-import ManageGroupSceneEntity
-
-public protocol DeleteGroupWorkable {
-  func deleteGroup(request: ManageGroup.DeleteGroup.Request)
-}
